requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.9.2
pydantic-settings==2.4.0
sqlalchemy[asyncio]==2.0.35
asyncpg==0.30.0
cryptography==43.0.1
redis[hiredis]==5.1.1
python-dotenv==1.0.1
alembic==1.13.2
structlog==24.4.0
pycryptodome==3.23.0
python-jose
celery[redis]
